fixing failing tests
